Combine StorageAuthScheme into SharedKeyAuthScheme
